Redux folder refactored. Fixed Edit audio cover screen Graident border not sync with current data (added sliderValue field to database).
